+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 4.1.0 - 2025-06-20
+
+### Added
+- Added daily voting limits functionality
+  - New 'daily' frequency option for contests
+  - Configurable reset time (defaults to midnight)
+  - Enforces one vote per user per day
+
 ## 1.0.0 - 2021-06-01
 
 ### Added